+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Notable updates:
 - Support for nv18.
 
 ### Added
+<<<<<<< HEAD
 
 - [database] added ParityDb statistics to the stats endpoint.
   [#2444](https://github.com/ChainSafe/forest/pull/2444)
@@ -21,7 +22,8 @@ Notable updates:
 - [forest daemon] Support for NV18.
   [#2558](https://github.com/ChainSafe/forest/pull/2558)
   [#2579](https://github.com/ChainSafe/forest/pull/2579)
-
+* [forest-cli info] Added forest info command. 
+  [#2578](https://github.com/ChainSafe/forest/pull/2578)
 ### Changed
 
 - [database] Move blockstore meta-data to standalone files.